Rename image_based_lighting_baker.h/.cpp to light_probe_baker.h/.cpp
