+++ projects/PathosEngine/src/pathos/render/light_probe_baker.h
