+++ projects/PathosEngine/src/pathos/render/sky_atmosphere.cpp
@@ -1,6 +1,6 @@
 #include "sky_atmosphere.h"
 #include "pathos/render/scene_render_targets.h"
-#include "pathos/render/image_based_lighting_baker.h"
+#include "pathos/render/light_probe_baker.h"
 #include "pathos/rhi/render_device.h"
 #include "pathos/rhi/shader_program.h"
 #include "pathos/rhi/texture.h"
